Address some of Fangrui's review comments
