+++ lld/ELF/Target.cpp
@@ -26,7 +26,6 @@
 #include "Target.h"
 #include "InputFiles.h"
 #include "OutputSections.h"
-#include "Relocations.h"
 #include "SymbolTable.h"
 #include "Symbols.h"
 #include "SyntheticSections.h"
